feat: Add algorithm header and implementation files
This commit adds the header file "algorithm.h" and the implementation file "algorithm.cpp" to the project. The "algorithm.h" file contains the declaration of the "sort" function, while the "algorithm.cpp" file defines the implementation of the "sort" function. These files are necessary for the project to compile and run successfully.